produre store

+++ demo/src/main/java/com/example/demo/repository/ItemTypeRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.ItemType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemTypeRepository extends JpaRepository<ItemType,Long> {
+public interface ItemTypeRepository extends JpaRepository<ItemType, Integer> {
 }
